requirements.txt
Flask~=2.3.2
Flask-Cors~=3.0.10
google-generativeai
python-dotenv
protobuf~=5.29.4